Feature/import latest

+++ scripts/import.ts
@@ -2,10 +2,28 @@ import { simpleGit, SimpleGit, CleanOptions } from 'simple-git';
 import * as fs from 'fs';
 import * as path from 'path';
 
-const git: SimpleGit = simpleGit().clean(CleanOptions.FORCE);
 const remoteUrl = 'https://github.com/Amsterdam/development-standards';
 const localDir = 'docs';
 const cloneDir = path.join(localDir, 'latest');
+
+async function cloneAndCheckout(repoUrl: string, branchName = 'main'): Promise<void> {
+  const git: SimpleGit = simpleGit().clean(CleanOptions.FORCE);
+
+  try {
+    // Clone the repository and change the current directory otherwise simple-git
+    // commands will run on this application's git config
+    await git.clone(repoUrl, cloneDir).cwd({ path: cloneDir });
+
+    // Change to the desired branch
+    await git.checkout(branchName);
+
+    cleanupRepoFiles();
+
+    console.log('\x1b[36m', 'Docs imported!');
+  } catch (error) {
+    console.error('Error occurred:', error);
+  }
+}
 
 // Cleanup previously cloned files from development-standards repo
 const cleanupOldFiles = () => {
@@ -21,21 +39,24 @@ const cleanupOldFiles = () => {
 
 // Sort and cleanup the cloned repository files
 const cleanupRepoFiles = () => {
-  const dirs = ['backend', 'cloud', 'frontend', 'standards'];
-  const renameDirs: { [key: string]: string } = { standards: 'general' };
+  const dirs = ['backend', 'cloud', 'frontend', 'general'];
+  // Below is empty but left in case a directory name requires changing on import
+  const renameDirs: { [key: string]: string } = {};
 
   // For each directory remove the README and move the directory up a level
   dirs.forEach((dir) => {
-    const readmePath = path.join(cloneDir, `${dir}/README.md`);
+    if (fs.existsSync(path.join(cloneDir, dir))) {
+      const readmePath = path.join(cloneDir, `${dir}/README.md`);
 
-    // We don't want any README.md files
-    if (fs.existsSync(readmePath)) {
-      fs.rmSync(path.join(cloneDir, `${dir}/README.md`));
+      // We don't want any README.md files
+      if (fs.existsSync(readmePath)) {
+        fs.rmSync(path.join(cloneDir, `${dir}/README.md`));
+      }
+
+      // TODO test markdown compilation for each file
+      const targetDir = renameDirs[dir] ?? dir;
+      fs.renameSync(path.join(cloneDir, dir), path.join(localDir, targetDir));
     }
-
-    // TODO test markdown compilation for each file
-    const targetDir = renameDirs[dir] ?? dir;
-    fs.renameSync(path.join(cloneDir, dir), path.join(localDir, targetDir));
   });
 
   // Erase repo directory
@@ -47,7 +68,4 @@ if (fs.existsSync(localDir)) {
 }
 
 // Clone the latest development-standards repo
-git
-  .clone(remoteUrl, cloneDir)
-  .then(() => cleanupRepoFiles())
-  .catch((err) => console.error('failed: ', err));
+cloneAndCheckout(remoteUrl, 'develop');
